Add files via upload
This excel will provide information on different types of bank transactions that are done. It will provide information of channels that are used for the transaction purpose like Credit or debit. 
Now,there are different channels of transaction that includes ATM, Online, Branch. 
The dataset will provide insights on the maximum and minimum transaction, choice of transaction based on profession, total transaction done for different types. 
The different tests used are Pivot table, conditional formatting, sorting, VLOOKUP, Pivot chart, filter.